Moved case style converters from Language to Conversions
Makes sense to put that logic there.
