+++ src/Conversions/Casing/FileSystemLowerCaseConverter.ts
