+++ src/Conversions/Casing/FileSystemUpperCaseConverter.ts
