+++ test/integration/FileEnd/file end.gls
@@ -1,5 +1,5 @@
 -
 file start : Program File
     comment line : ...
-file end : File
+file end
 -
