+++ test/unit/Conversions/Casing/CamelCaseConverterTests.ts
@@ -1,7 +1,7 @@
 import { expect } from "chai";
 import "mocha";
 
-import { CamelCaseConverter } from "../../../../lib/Languages/Casing/CamelCaseConverter";
+import { CamelCaseConverter } from "../../../../lib/Conversions/Casing/CamelCaseConverter";
 import { itConvertsFromTo } from "./ConverterTests";
 
 describe("CamelCaseConverter", () => {
