+++ test/unit/Conversions/Casing/PascalCaseConverterTests.ts
@@ -1,7 +1,7 @@
 import { expect } from "chai";
 import "mocha";
 
-import { PascalCaseConverter } from "../../../../lib/Languages/Casing/PascalCaseConverter";
+import { PascalCaseConverter } from "../../../../lib/Conversions/Casing/PascalCaseConverter";
 import { itConvertsFromTo } from "./ConverterTests";
 
 describe("PascalCaseConverter", () => {
