+++ test/unit/Conversions/Casing/SnakeCaseConverterTests.ts
@@ -1,7 +1,7 @@
 import { expect } from "chai";
 import "mocha";
 
-import { SnakeCaseConverter } from "../../../../lib/Languages/Casing/SnakeCaseConverter";
+import { SnakeCaseConverter } from "../../../../lib/Conversions/Casing/SnakeCaseConverter";
 import { itConvertsFromTo } from "./ConverterTests";
 
 describe("SnakeCaseConverter", () => {
